API Reference

API Reference for Swarm.

Swarm.Agent

Swarm.Response

Swarm.Session

Swarm.Session

Swarm.add_tools!

Swarm.handle_tool_calls!

Swarm.run_full_turn

Swarm.update_system_message!

Swarm.Agent — Type

Agent

Agent is a stateless struct that holds the the reference to LLM, tools and the instructions.

Fields

name::String: The name of the agent.

model::String: The model to use for the agent.

instructions::String: The instructions for the agent.

tool_map::Dict{String, AbstractTool}: A dictionary of tools available to the agent.

tool_choice::Union{String, Nothing}: The tool choice for the agent.

parallel_tool_calls::Bool: Whether to allow parallel tool calls. Defaults to true - NOT SUPPORTED YET.

Swarm.Response — Type

Response

Response from a single full turn of swarm.

Fields

messages::Vector{PT.AbstractMessage}: The additional messages from the last full turn.

agent::Union{Agent, Nothing}: The current active agent in the session.

context::Dict{Symbol, Any}: The context variables or other data in the session.

Swarm.Session — Type

Session

Session is a mutable struct that holds the messages, agent and context.

Fields

messages::Vector{PT.AbstractMessage}: The history of chat or tool messages in the session.

agent::Union{Agent, Nothing}: The current active agent in the session.

context::Dict{Symbol, Any}: The context variables or other data in the session.

Swarm.Session — Method

Session(agent::Agent;
    context::Dict{Symbol, Any} = Dict{Symbol, Any}())

Initializes a Session with an agent and an optional context.

Run run_full_turn! with a user_prompt to continue the session.

Swarm.add_tools! — Method

add_tools!(agent::Agent, tools::Vector)

Adds tools to an agent.

Swarm.handle_tool_calls! — Function

handle_tool_calls!(
    active_agent::Union{Agent, Nothing}, history::AbstractVector{<:PT.AbstractMessage},
    context::Dict{Symbol, Any} = Dict{Symbol, Any}())

Handles tool calls for an agent.

Swarm.run_full_turn — Function

run_full_turn(agent::Agent, messages::AbstractVector{<:PT.AbstractMessage},
    context::Dict{Symbol, Any} = Dict{Symbol, Any}(); max_turns::Int = 5,
    kwargs...)

Runs a full turn of an agent (executes all tool calls).

Swarm.update_system_message! — Method

update_system_message!(history::AbstractVector{<:PT.AbstractMessage},
    active_agent::Union{Agent, Nothing})

Updates the system message in the history (1st message) with the active agent's instructions.